Unify user and account models
