+++ cooking_core/accounts/admin.py
@@ -1,5 +1,5 @@
 from django.contrib import admin
 
-from .models import User
+from .models import Account
 
-admin.site.register(User)
+admin.site.register(Account)
