+++ cooking_core/accounts/models/__init__.py
@@ -1,2 +1,1 @@
 from .account import Account  # noqa: F401
-from .user import User  # noqa: F401
